Refactor theme provider and set theme on :root
