+++ apps/web/src/App.tsx
@@ -2,11 +2,13 @@ import { useMemo, useState } from "react";
 import { createApiClient, ApiClientProvider } from "@yas/api-client";
 import { BrowserRouter } from "react-router-dom";
 import { ErrorBoundary } from "react-error-boundary";
+import { ThemeProvider, ThemeInjector } from "@yas/ui";
+import { getPreferredTheme } from "@yas/ui";
 import { env } from "./env";
 import { AppRoutes } from "./Routes";
-import { ThemeProvider } from "./theme/ThemeProvider";
-import { getPreferredTheme } from "./theme";
 import { ErrorFallback } from "./components/ErrorFallback";
+
+const rootRef = { current: document.documentElement };
 
 export default function App() {
   const [theme, setTheme] = useState(getPreferredTheme);
@@ -17,6 +19,7 @@ export default function App() {
       <BrowserRouter>
         <ApiClientProvider value={apiClient}>
           <ThemeProvider theme={theme} setTheme={setTheme}>
+            <ThemeInjector target={rootRef} />
             <AppRoutes />
           </ThemeProvider>
         </ApiClientProvider>
