+++ apps/web/src/components/Layout.tsx
@@ -1,15 +1,12 @@
 import { CircularProgress, Container, ModalOutlet, Dock, Stack } from "@yas/ui";
 import { Outlet as RouterOutlet } from "react-router-dom";
-import { clsx } from "@yas/style";
 import { Suspense } from "react";
-import { useThemeClassName } from "../theme";
 import { Menu } from "./Menu";
 import { layout } from "./Layout.css";
 
 export default function Layout() {
-  const theme = useThemeClassName();
   return (
-    <div className={clsx(theme, layout)}>
+    <div className={layout}>
       <ContainerStack>
         <Menu />
         <Stack sx={{ flex: 1 }}>
