+++ libraries/yas-ui/src/index.ts
@@ -36,3 +36,4 @@ export * from "./hooks/useElementBounds";
 export * from "./hooks/useModal";
 export * from "./hooks/useMediaQuery";
 export * from "./debug/ExampleImage";
+export * from "./ThemeProvider";
